always export objects, merge layer files
